Move `enum Command` from `device::trace` to `command::encoder_command`.
