+++ wgpu-core/src/command/render.rs
@@ -1753,7 +1753,7 @@ impl Global {
             let cmd_buf_data = cmd_buf_data.get_inner();
 
             if let Some(ref mut list) = cmd_buf_data.commands {
-                list.push(crate::device::trace::Command::RunRenderPass {
+                list.push(crate::command::Command::RunRenderPass {
                     base: BasePass {
                         label: base.label.clone(),
                         error: None,
